Break each class into separate file, add private section for the openmp parallelization
