+++ CMakeLists.txt
@@ -29,12 +29,12 @@ if (CUDA)
   add_definitions(-std=c++11 -DCUDA_ENABLED=1)
   list(APPEND CUDA_NVCC_FLAGS "-O3;-std=c++11")
   SET(CUDA_PROPAGATE_HOST_FLAGS OFF)
-  CUDA_ADD_EXECUTABLE(peticodiac src/main.cpp src/CpuSolver.cpp src/CudaSolver.cu src/generalSimplex.cu src/Solver.cpp)
+  CUDA_ADD_EXECUTABLE(peticodiac src/main.cpp src/CpuSolver.cpp src/CpuEagerSolver.cpp src/Solver.cpp src/AbstractSolver.cpp src/CudaSolver.cu src/generalSimplex.cu)
 else()
   set(GCC_COVERAGE_COMPILE_FLAGS "${GCC_COVERAGE_COMPILE_FLAGS} -Wall -std=c++11 -fopenmp")
   set(GCC_EXE_LINKER_FLAGS "-fopenmp")
 
   set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} ${GCC_COVERAGE_COMPILE_FLAGS}")
   set(CMAKE_EXE_LINKER_FLAGS "${CMAKE_EXE_LINKER_FLAGS} ${GCC_EXE_LINKER_FLAGS}")
-  add_executable(peticodiac src/main.cpp src/CpuSolver.cpp src/Solver.cpp)
+  add_executable(peticodiac src/main.cpp src/CpuSolver.cpp src/CpuEagerSolver.cpp src/Solver.cpp src/AbstractSolver.cpp src/SolverProfiler.cpp)
 endif()
